# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Optional role header for consistency. Auth endpoints do not enforce authentication.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Validate seeded credentials and return the user. No auth token is required.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Optional role header for consistency. Auth endpoints do not enforce authentication.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Register a new organization with a COO account and initial departments.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Optional role header for consistency. Auth endpoints do not enforce authentication.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Request a password reset OTP for the given email. Returns simulated OTP for project purposes.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Optional role header for consistency. Auth endpoints do not enforce authentication.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Reset account password using the OTP received from forgot-password.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/alerts": {
      "get": {
        "operationId": "AlertsController_list",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Manager/COO respond to alerts; Analyst/Super User can inspect.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "alerts"
        ]
      },
      "post": {
        "operationId": "AlertsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Manager/COO respond to alerts; Analyst/Super User can inspect.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/{id}": {
      "get": {
        "operationId": "AlertsController_find",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Manager/COO respond to alerts; Analyst/Super User can inspect.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "alerts"
        ]
      },
      "patch": {
        "operationId": "AlertsController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Manager/COO respond to alerts; Analyst/Super User can inspect.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "alerts"
        ]
      },
      "put": {
        "operationId": "AlertsController_replace",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Manager/COO respond to alerts; Analyst/Super User can inspect.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "alerts"
        ]
      },
      "delete": {
        "operationId": "AlertsController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Manager/COO respond to alerts; Analyst/Super User can inspect.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/{id}/respond": {
      "post": {
        "operationId": "AlertsController_respond",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Manager/COO respond to alerts; Analyst/Super User can inspect.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Resolve an alert with business response text.",
        "tags": [
          "alerts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rorizon Review-4 API",
    "description": "NestJS in-memory backend with RBAC role headers for Rorizon workflows.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role for RBAC authorization. Allowed values: Super User, COO, Manager, Analyst."
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@platformops.com"
          },
          "password": {
            "type": "string",
            "example": "Admin@123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "coo@company.com",
            "description": "COO email address"
          },
          "password": {
            "type": "string",
            "example": "Secure@123",
            "description": "Minimum 8 characters"
          },
          "organizationName": {
            "type": "string",
            "example": "Acme Industries",
            "description": "Organization name"
          },
          "cooName": {
            "type": "string",
            "example": "Jane Smith",
            "description": "COO full name"
          },
          "departments": {
            "type": "string",
            "example": "Operations, Manufacturing",
            "description": "Comma-separated department names"
          }
        },
        "required": [
          "email",
          "password",
          "organizationName",
          "cooName"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@company.com",
            "description": "Email of the account to reset"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@company.com"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "OTP code received"
          },
          "newPassword": {
            "type": "string",
            "example": "NewSecure@123",
            "description": "New password (minimum 8 characters)"
          }
        },
        "required": [
          "email",
          "otp",
          "newPassword"
        ]
      },
...
      "AlertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "alt-003"
          },
          "submissionId": {
            "type": "string",
            "example": "sub-001"
          },
          "title": {
            "type": "string",
            "example": "Submission Correction Required"
          },
          "type": {
            "type": "string",
            "example": "Threshold Breach"
          },
          "severity": {
            "type": "string",
            "example": "Critical"
          },
          "organizationId": {
            "type": "string",
            "example": "org-techcorp"
          },
          "departmentId": {
            "type": "string",
            "example": "dept-mfg"
          },
          "departmentName": {
            "type": "string",
            "example": "Manufacturing"
          },
          "status": {
            "type": "string",
            "example": "Open"
          },
          "message": {
            "type": "string",
            "example": "Consumption exceeded threshold."
          },
          "roleScope": {
            "example": [
              "Manager",
              "COO"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "period": {
            "type": "string",
            "example": "03 2026"
          },
          "deviationReason": {
            "type": "string",
            "example": "Temporary generator usage."
          },
          "managerExplanation": {
            "type": "string",
            "example": "Generator test conducted."
          },
          "managerActionTaken": {
            "type": "string",
            "example": "Reduced generator hours."
          },
          "managerName": {
            "type": "string",
            "example": "Robert Johnson"
          },
          "cooDirective": {
            "type": "string",
            "example": "Limit generator operation to peak hours."
          },
          "analystClarificationNotes": {
            "type": "string",
            "example": "Provide generator log files."
          },
          "response": {
            "type": "string",
            "example": "Corrective action completed."
          },
          "resolvedAt": {
            "type": "string",
            "example": "Just now"
          },
          "resolvedBy": {
            "type": "string",
            "example": "John Anderson"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-03-07 14:00:00"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-03-07 14:00:00"
          }
        },
        "required": [
          "type",
          "severity",
          "departmentId",
          "status",
          "message"
        ]
      },
      "UpdateAlertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "alt-003"
          },
          "submissionId": {
            "type": "string",
            "example": "sub-001"
          },
          "title": {
            "type": "string",
            "example": "Submission Correction Required"
          },
          "type": {
            "type": "string",
            "example": "Threshold Breach"
          },
          "severity": {
            "type": "string",
            "example": "Critical"
          },
          "organizationId": {
            "type": "string",
            "example": "org-techcorp"
          },
          "departmentId": {
            "type": "string",
            "example": "dept-mfg"
          },
          "departmentName": {
            "type": "string",
            "example": "Manufacturing"
          },
          "status": {
            "type": "string",
            "example": "Open"
          },
          "message": {
            "type": "string",
            "example": "Consumption exceeded threshold."
          },
          "roleScope": {
            "example": [
              "Manager",
              "COO"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "period": {
            "type": "string",
            "example": "03 2026"
          },
          "deviationReason": {
            "type": "string",
            "example": "Temporary generator usage."
          },
          "managerExplanation": {
            "type": "string",
            "example": "Generator test conducted."
          },
          "managerActionTaken": {
            "type": "string",
            "example": "Reduced generator hours."
          },
          "managerName": {
            "type": "string",
            "example": "Robert Johnson"
          },
          "cooDirective": {
            "type": "string",
            "example": "Limit generator operation to peak hours."
          },
          "analystClarificationNotes": {
            "type": "string",
            "example": "Provide generator log files."
          },
          "response": {
            "type": "string",
            "example": "Corrective action completed."
          },
          "resolvedAt": {
            "type": "string",
            "example": "Just now"
          },
          "resolvedBy": {
            "type": "string",
            "example": "John Anderson"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-03-07 14:00:00"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-03-07 14:00:00"
          }
        }
      },
      "AlertResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "string",
            "example": "Corrective action completed."
          }
        },
        "required": [
          "response"
        ]
      },
...
    }
  }
}